Implement argSort, argMax, argMin, swapAxes, transpose
change the type of the axis from int to long

+++ api/src/main/java/ai/djl/ndarray/NDArray.java
@@ -27,6 +27,7 @@ import java.nio.LongBuffer;
 import java.util.Arrays;
 import java.util.function.Predicate;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 /**
  * An interface representing an n-dimensional array.
@@ -2919,7 +2920,7 @@ public interface NDArray extends AutoCloseable {
      * @return the result {@code NDArray}. The number of dimensions is one greater than that of the
      *     {@code NDArray}
      */
-    NDArray expandDims(int axis);
+    NDArray expandDims(long axis);
 
     /**
      * Removes all singleton dimensions from this {@code NDArray} {@link Shape}.
@@ -2944,7 +2945,10 @@ public interface NDArray extends AutoCloseable {
      */
     default NDArray squeeze() {
         long[] shape = getShape().getShape();
-        return squeeze(IntStream.range(0, shape.length).filter(i -> shape[i] == 1).toArray());
+        return squeeze(
+                LongStream.range(0, shape.length)
+                        .filter(i -> shape[Math.toIntExact(i)] == 1)
+                        .toArray());
     }
 
     /**
@@ -2977,8 +2981,8 @@ public interface NDArray extends AutoCloseable {
      * @return a result {@code NDArray} of same size and data without the axis at part of the shape
      * @throws IllegalArgumentException thrown if the given axis is not a singleton dimension
      */
-    default NDArray squeeze(int axis) {
-        return squeeze(new int[] {axis});
+    default NDArray squeeze(long axis) {
+        return squeeze(new long[] {axis});
     }
 
     /**
@@ -3005,7 +3009,7 @@ public interface NDArray extends AutoCloseable {
      * @throws IllegalArgumentException thrown if any of the given axes are not a singleton
      *     dimension
      */
-    NDArray squeeze(int[] axes);
+    NDArray squeeze(long[] axes);
 
     /**
      * Joins a {@code NDArray} along the first axis.
@@ -3258,7 +3262,7 @@ public interface NDArray extends AutoCloseable {
      *
      * @return a {@code NDArray} of indices corresponding to elements in this {@code NDArray} on the
      *     axis, the output DataType is always {@link DataType#INT32}
-     * @see NDArray#argSort(int, boolean)
+     * @see NDArray#argSort(long, boolean)
      */
     default NDArray argSort() {
         return argSort(-1, true);
@@ -3289,7 +3293,7 @@ public interface NDArray extends AutoCloseable {
      * @param axis the axis to sort along
      * @return a {@code NDArray} of indices corresponding to elements in this {@code NDArray} on the
      *     axis, the output DataType is always {@link DataType#INT32}
-     * @see NDArray#argSort(int, boolean)
+     * @see NDArray#argSort(long, boolean)
      */
     default NDArray argSort(int axis) {
         return argSort(axis, true);
@@ -3317,7 +3321,7 @@ public interface NDArray extends AutoCloseable {
      * @return a {@code NDArray} of indices corresponding to elements in this {@code NDArray} on the
      *     axis, the output DataType is always {@link DataType#INT32}
      */
-    NDArray argSort(int axis, boolean ascending);
+    NDArray argSort(long axis, boolean ascending);
 
     /**
      * Sorts the flattened {@code NDArray}.
@@ -3372,10 +3376,10 @@ public interface NDArray extends AutoCloseable {
      * @param axis the axis along which to apply
      * @return the result {@code NDArray}
      * @see <a href="https://en.wikipedia.org/wiki/Softmax_function">softmax</a>
-     * @see NDArray#softmax(int[], double)
+     * @see NDArray#softmax(long[], double)
      */
-    default NDArray softmax(int axis) {
-        return softmax(new int[] {axis}, 1);
+    default NDArray softmax(long axis) {
+        return softmax(new long[] {axis}, 1);
     }
 
     /**
@@ -3385,10 +3389,10 @@ public interface NDArray extends AutoCloseable {
      * @param temperature the exponent multiplier Beta in the softmax
      * @return the result {@code NDArray}
      * @see <a href="https://en.wikipedia.org/wiki/Softmax_function">softmax</a>
-     * @see NDArray#softmax(int[], double)
+     * @see NDArray#softmax(long[], double)
      */
-    default NDArray softmax(int axis, double temperature) {
-        return softmax(new int[] {axis}, temperature);
+    default NDArray softmax(long axis, double temperature) {
+        return softmax(new long[] {axis}, temperature);
     }
 
     /**
@@ -3399,7 +3403,7 @@ public interface NDArray extends AutoCloseable {
      * @return the result {@code NDArray}
      * @see <a href="https://en.wikipedia.org/wiki/Softmax_function">softmax</a>
      */
-    default NDArray softmax(int[] axes) {
+    default NDArray softmax(long[] axes) {
         return softmax(axes, 1);
     }
 
@@ -3411,9 +3415,9 @@ public interface NDArray extends AutoCloseable {
      * @param temperature the exponent multiplier Beta in the softmax
      * @return the result {@code NDArray}
      * @see <a href="https://en.wikipedia.org/wiki/Softmax_function">softmax</a>
-     * @see NDArray#softmax(int[], double)
+     * @see NDArray#softmax(long[], double)
      */
-    NDArray softmax(int[] axes, double temperature);
+    NDArray softmax(long[] axes, double temperature);
 
     /**
      * Applies the softmax function followed by a logarithm.
@@ -3424,8 +3428,8 @@ public interface NDArray extends AutoCloseable {
      * @param axis the axis along which to apply
      * @return the result {@code NDArray}
      */
-    default NDArray logSoftmax(int axis) {
-        return logSoftmax(new int[] {axis}, 1);
+    default NDArray logSoftmax(long axis) {
+        return logSoftmax(new long[] {axis}, 1);
     }
 
     /**
@@ -3438,8 +3442,8 @@ public interface NDArray extends AutoCloseable {
      * @param temperature the exponent multiplier Beta in the softmax
      * @return the result {@code NDArray}
      */
-    default NDArray logSoftmax(int axis, double temperature) {
-        return logSoftmax(new int[] {axis}, temperature);
+    default NDArray logSoftmax(long axis, double temperature) {
+        return logSoftmax(new long[] {axis}, temperature);
     }
 
     /**
@@ -3452,7 +3456,7 @@ public interface NDArray extends AutoCloseable {
      *     the whole array.
      * @return the result {@code NDArray}
      */
-    default NDArray logSoftmax(int[] axes) {
+    default NDArray logSoftmax(long[] axes) {
         return logSoftmax(axes, 1);
     }
 
@@ -3467,7 +3471,7 @@ public interface NDArray extends AutoCloseable {
      * @param temperature the exponent multiplier Beta in the softmax
      * @return the result {@code NDArray}
      */
-    NDArray logSoftmax(int[] axes, double temperature);
+    NDArray logSoftmax(long[] axes, double temperature);
 
     /**
      * Returns the cumulative sum of the elements in the flattened {@code NDArray}.
@@ -3826,11 +3830,11 @@ public interface NDArray extends AutoCloseable {
      * @param axis2 the second axis
      * @return the swapped axes {@code NDArray}
      */
-    default NDArray swapAxes(int axis1, int axis2) {
-        int[] dims = IntStream.range(0, getShape().dimension()).toArray();
-        int tmp = dims[axis1];
-        dims[axis1] = dims[axis2];
-        dims[axis2] = tmp;
+    default NDArray swapAxes(long axis1, long axis2) {
+        long[] dims = LongStream.range(0, getShape().dimension()).toArray();
+        long tmp = dims[Math.toIntExact(axis1)];
+        dims[Math.toIntExact(axis1)] = dims[Math.toIntExact(axis2)];
+        dims[Math.toIntExact(axis2)] = tmp;
         return transpose(dims);
     }
 
@@ -3900,7 +3904,7 @@ public interface NDArray extends AutoCloseable {
      * @throws IllegalArgumentException thrown when passing a axis that is greater than the actual
      *     number of dimensions
      */
-    NDArray transpose(int... axes);
+    NDArray transpose(long... axes);
 
     /**
      * Broadcasts this {@code NDArray} to be the given shape.
@@ -4004,7 +4008,7 @@ public interface NDArray extends AutoCloseable {
      * @param axis the axis along which to find maximum values
      * @return a {@code NDArray} containing indices
      */
-    NDArray argMax(int axis);
+    NDArray argMax(long axis);
 
     /**
      * Returns the indices of the minimum values into the flattened {@code NDArray}.
@@ -4050,7 +4054,7 @@ public interface NDArray extends AutoCloseable {
      * @param axis the axis along which to find minimum values
      * @return a {@code NDArray} containing indices
      */
-    NDArray argMin(int axis);
+    NDArray argMin(long axis);
 
     /**
      * Returns percentile for this {@code NDArray}.
