+++ api/src/main/java/ai/djl/ndarray/types/Shape.java
@@ -140,16 +140,16 @@ public class Shape {
      * @return the size of specific dimension(s) or -1 for indeterminate size
      * @throws IllegalArgumentException thrown if passed an invalid dimension
      */
-    public long size(int... dimensions) {
+    public long size(long... dimensions) {
         int total = 1;
-        for (int d : dimensions) {
+        for (long d : dimensions) {
             if (d < 0 || d >= shape.length) {
                 throw new IllegalArgumentException("Invalid dimension " + d);
             }
-            if (shape[d] == -1) {
+            if (shape[Math.toIntExact(d)] == -1) {
                 return -1;
             }
-            total *= shape[d];
+            total *= shape[Math.toIntExact(d)];
         }
         return total;
     }
