+++ tensorflow/engine/src/main/java/ai/djl/tensorflow/engine/TfNDArray.java
@@ -780,13 +780,13 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray expandDims(int axis) {
+    public NDArray expandDims(long axis) {
         return null;
     }
 
     /** {@inheritDoc} */
     @Override
-    public NDArray squeeze(int[] axes) {
+    public NDArray squeeze(long[] axes) {
         return null;
     }
 
@@ -816,7 +816,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argSort(int axis, boolean ascending) {
+    public NDArray argSort(long axis, boolean ascending) {
         return null;
     }
 
@@ -834,7 +834,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray softmax(int[] axes, double temperature) {
+    public NDArray softmax(long[] axes, double temperature) {
         Operation op =
                 manager.getGraph()
                         .opBuilder("Softmax", "Softmax_" + TfNDManager.nextNameAssignment())
@@ -859,7 +859,7 @@ public class TfNDArray implements NDArray {
     }
 
     @Override
-    public NDArray logSoftmax(int[] axes, double temperature) {
+    public NDArray logSoftmax(long[] axes, double temperature) {
         return null;
     }
 
@@ -961,7 +961,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray swapAxes(int axis1, int axis2) {
+    public NDArray swapAxes(long axis1, long axis2) {
         return null;
     }
 
@@ -973,7 +973,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray transpose(int... dimensions) {
+    public NDArray transpose(long... dimensions) {
         return null;
     }
 
@@ -997,7 +997,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argMax(int axis) {
+    public NDArray argMax(long axis) {
         return null;
     }
 
@@ -1009,7 +1009,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argMin(int axis) {
+    public NDArray argMin(long axis) {
         return null;
     }
